Move file watcher logics into separate file (`key_source_file.go`)
No functional behavior changes. Includes unit tests.
